Remove C++ compilation dependencies from base
These are currently needed only for compiling Greenbar. However, they take up about 150MB of space. Due to the nature of Docker's layering system, that means that we'll need to carry that around for all images that depend on our base! This almost doubles the size of our production Cog image, which is unacceptable.

Since we don't care so much about size for testing images, we can simply add the relevant package additions when we need them for tests.