Create an Elxiir container based on Debian
This is only to support CI builds of Cog Console that require NodeJS
and PhantomJS. Getting those to run in our normal Alpine-based
containers is a royal pain. It was easier to just build a Debian-based
(more importantly, glibc-based) container and use that.

It does mean that we have two places to update when we migrate to
newer versions of Erlang / Elixir, but that's not too onerous right
now.